apps/python/python-requirements-manager/requirements.txt
# System and Utility
psutil==5.9.8                 # CPU/memory monitoring
pyyaml==6.0.1                 # YAML config parsing
numpy==1.26.4                 # Numerical computing
python-dotenv==1.0.0          # Environment variables

# Web & API
flask==2.3.3                  # REST API server
authlib==1.3.1                # OAuth integration with Flask
requests==2.31.0              # General HTTP requests
urllib3==2.0.7                # HTTP client
werkzeug==2.3.7               # WSGI utilities

# Cloud SDKs
boto3==1.34.112               # AWS SDK (S3 sync)
botocore==1.34.112            # AWS core library
google-auth==2.29.0           # Google authentication
google-api-python-client==2.126.0 # Google Drive API support
google-auth-oauthlib==1.2.0   # Google OAuth
google-auth-httplib2==0.2.0   # Google HTTP transport

# Background Tasks and Messaging
redis==5.0.1                  # Pub/Sub and task distribution
celery==5.3.4                 # Distributed task queue
kombu==5.3.4                  # Messaging library

# ML and Automation
scikit-learn==1.5.0           # AI priority model
joblib==1.4.2                 # Model persistence (used by sklearn)
threadpoolctl==3.5.0          # Low-level thread control (used by sklearn)
pandas==2.0.3                 # Data manipulation
matplotlib==3.7.2             # Plotting

# Cryptography and Security
cryptography==42.0.5          # Fernet encryption
pycryptodome==3.20.0          # Optional extended crypto support
bcrypt==4.1.2                 # Password hashing
jwt==1.3.1                    # JSON Web Tokens

# Emulator / VM Layer
unicorn==2.0.1                # CPU emulator (X86)
capstone==5.0.1               # Disassembler (optional but often paired)
keystone-engine==0.9.2        # Assembler engine

# Imaging and Snapshots
pillow==10.3.0                # For snapshot PNG generation
opencv-python==4.8.1.78       # Computer vision

# Database
sqlalchemy==2.0.23            # ORM
alembic==1.13.1               # Database migrations

# Testing Frameworks
pytest==8.2.1                 # Unit testing
pytest-asyncio==0.23.6        # Async test support
coverage==7.4.4               # Coverage reporting
pytest-cov==4.1.0             # Coverage plugin

# Build and Packaging
setuptools==70.0.0            # Core packaging utility
wheel==0.43.0                 # Build wheels for distribution
build==1.0.3                  # PEP 517 build frontend

# Optional Dev Tools
black==24.4.2                 # Code formatter
mypy==1.10.0                  # Type checking
isort==5.13.2                 # Import sorting
flake8==7.0.0                 # Linting
pre-commit==3.6.0             # Git hooks

# Compatibility / Safety
importlib-metadata==7.1.0     # Backport for metadata
typing-extensions==4.12.0     # Typing backports
packaging==23.2               # Core packaging utilities
six==1.16.0                   # Python 2/3 compatibility

# Additional utilities
click==8.1.7                  # CLI framework
rich==13.7.0                  # Rich text and beautiful formatting
loguru==0.7.2                 # Better logging
tenacity==8.2.3               # Retry library